requirements.txt
torch==2.3.1
torchvision==0.18.1
xformers==0.0.27
accelerate==0.32.0
diffusers>=0.30.0
transformers==4.37.2
safetensors==0.4.2
peft==0.7.0
huggingface_hub==0.23.4
einops==0.8.0
timm==0.6.12
lpips==0.1.4
brisque==0.0.17
scikit-image==0.24.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
scipy==1.14.0
Pillow==10.3.0
matplotlib==3.9.1
tqdm==4.66.4
pyyaml==6.0.1
requests==2.32.3
rich==13.9.3
tensorboard==2.17.0
wandb==0.17.4
git+https://github.com/openai/CLIP.git